corrected suggested improvements
fixed
