--- src/sprintThree/Cars/oopExercises.java
@@ -1,16 +1,0 @@
-//OOP Exercises.
-
-package sprintThree.Cars;
-
-//Part 1
-//Create Car class.
-//Create following attributes: //Country of car //Model //Maker
-
-class Car {
-    //attributes
-    String country;
-    String model;
-    String maker;
-}
-
-
